Create Pizza_Runner_Case_Study_Solution
This case study is divided into two areas of focus: Pizza Metrics and Runner and Customer Experience. It contains numerous questions that will require SQL queries to answer. However, before delving into the queries, it is essential to investigate the data and address the null values and data types present in the customer_orders and runner_orders tables.